fix: vram precision, max_days field name, status width, tighter columns
  vram_gb is now f32 (was u32) so 96.0 GB cards render as 96.0 not 95. The vast.ai field for max rental
  duration is duration (with max_days_in_use as deserialisation alias). Status column widened 8→10 to fit
  unverified. All perf+infra columns tightened to vastai's widths to fit a half-screen 16" MBP terminal.

+++ src/providers/mod.rs
@@ -20,7 +20,7 @@ pub struct Offer {
     pub id: String,
     pub gpu_name: String,
     pub num_gpus: u32,
-    pub vram_gb: u32,
+    pub vram_gb: f32,
     pub disk_gb: u32,
     pub price_per_hour_usd: f32,
     pub region: Option<String>,
